Add evaluation script for DenseSLM4 using LightEval
- Implemented `test_lighteval.py` to evaluate DenseSLM4 models with the LightEval library.
- Included functions to load the model and configuration, set up evaluation parameters, and run evaluations on specified tasks.
- Added command-line argument parsing for flexible execution.

+++ scripts/test_generation.py
@@ -34,7 +34,7 @@ def generate_from_checkpoint(checkpoint_path: str, prompt: str, max_new_tokens: 
     # Retie weights after loading
     model._retie_weights()
     
-    model = model.cuda()
+    model = model.cuda().to(torch.bfloat16)
     model.eval()
     print(f"Model loaded successfully!")
